move TemplateGuesser to related Helper dir
